refactor: rename tasks route to task tab
